+++ R/zizLaplacePosteriorEngine.R
@@ -58,6 +58,8 @@ fitLaplacePosteriorModel.zizModel = function(model,
     stop("prior must be an object of class psPrior")
   }
 
+  validateZetaPriorRange(prior$range)
+
   modelObservationData(model, x)
   approximation = makeZizPosteriorLaplace(
     x = x,