+++ R/zizNumericalPosterior.R
@@ -33,6 +33,8 @@ fitNumericalPosteriorModel.zizModel = function(model,
     stop("prior must be an object of class psPrior")
   }
 
+  validateZetaPriorRange(prior$range)
+
   modelObservationData(model, x)
   posteriorGrid = makeZizPosteriorGrid(
     x = x,